requirements.txt
flask
langchain
faiss-cpu
pdfplumber
openai
sentence-transformers
fastapi
uvicorn
PyMuPDF
python-dotenv